Change to pom for core change. Fix for Region switch bug. Expansion of alterVM.
Changed the pom.xml to represent the version of the current build.

Removed the region specification when doing listVLans as it seemed to
be causing a bug in which only one region would be recognised at a time
(the other region's resources were being lost by enStratus).

Expanded the alterVirtualMachine functionality to include Volume
additions.